refactor: rename CheckBox to ToolList and simplify CategoryBar category selection logic
