+++ __tests__/CheckBox.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, waitFor, act } from "@testing-library/react";
-import CheckBox from "@/app/toolkit/components/CheckBox";
+import CheckBox from "@/app/toolkit/components/ToolList";
 import DatabaseManager from "@/lib/db/DatabaseManager";
 //import { DndContext } from "@dnd-kit/core";
